Set paths for ACL usage
Given the default path of `hubot"storage` is under the root, if using ACLs,
the policy would need to be set to allow write permissions to root (/).

By setting it to `hubot/hubot:storage`, if not using ACLs, it would still
be created automatically but also allow you to set a policy as described
in README.md locking down access to only `hubot/`.
+